Create Proper File Structure for MVVM
