+++ View/CustomDialogBox.cs
