+++ View/CustomDialogBox.resx
